The MD5 utility class computes the MD5 digest (aka: "hash") of a block of data; an MD5 digest is a 32-char ASCII string.
The synchronized/static function "Digest" is useful for situations where
lock contention in the application is not expected to be an issue.

The unsynchronized/non-static method "digest" is useful in a
multi-threaded program that wanted to avoid locking by creating
an MD5 object for exclusive use by a single thread.




git-svn-id: svn://svn.alfresco.com/alfresco/HEAD@5552 c73165b8-c2a3-4c3a-8162-ddfb506fd858