YUI fixes:
Uploader.as - if a file with 0b as size is selected flash movie diesnt throw error
flash-upload.js - If a file with 0b as size is selected it is ignoerd but the user is warned about it
uploader-exprimental-xxx.js - Didn't accpet params as an argument which made it impossible to set param "wmode" to "transparent" which is necessary if the flashmove should obey the z-index and not hide parts of popups to be displayed.

Note! YUI is overriden by copying the web/yui folder in the end of the deploy script, yui is only overriden by slingshot.

git-svn-id: svn://svn.alfresco.com/alfresco/HEAD@10559 c73165b8-c2a3-4c3a-8162-ddfb506fd858