Moved WEF to a separate project and added targets to build the various artifacts required for AWE.
Also added a new macro to minimise JavaScript files for a given folder tree. Target is called 'minimise-javascript' and takes 2 parameters, 'sourcedir' and 'destdir'.

NOTE: The WEF project produces a surf JAR that can be deployed, but AWE is NOT currently using it. Further work is also needed in the target that builds AWE surf JAR file.

git-svn-id: svn://svn.alfresco.com/alfresco/HEAD@18816 c73165b8-c2a3-4c3a-8162-ddfb506fd858