Fix for ALF-2236: Does not display tabs for contents created at the forms32
A majority of the examples that were in the "ttl-forms-32.amp.sample" file, originally written by Yong, have been included in the FDK. The other forms examples present in the slingshot web-extension folder have also been added to the FDK, it now provides working examples of these and the examples described in http://wiki.alfresco.com/wiki/Forms_Examples.

Configuration has also been added to allow the creation of the custom types provided.

git-svn-id: svn://svn.alfresco.com/alfresco/HEAD@19781 c73165b8-c2a3-4c3a-8162-ddfb506fd858